[#4]Refactor: 소개 페이지 컴포넌트 이동
