+++ src/pages/about/components/ProfileCard/ProfileDescription.tsx
@@ -1,5 +1,4 @@
-import Image from "next/image";
-import styled from "styled-components";
+import styled from 'styled-components';
 
 const ProfileDescriptionBox = styled.article`
   display: flex;
@@ -14,7 +13,7 @@ const ProfileDescriptionBox = styled.article`
   }
 
   em {
-    font-weight: 700;;
+    font-weight: 700;
   }
 
   p {
@@ -26,27 +25,22 @@ const ProfileDescriptionBox = styled.article`
   p:last-child {
     margin-bottom: 0;
   }
-`
-
+`;
 
 const ProfileDescription = () => {
   return (
     <ProfileDescriptionBox>
       <h2>
-        안녕하세요, <br/>
+        안녕하세요, <br />
         프론트엔드 지망생 <em>김민규</em>입니다.
       </h2>
-      <p>
-        몰입과 경험을 중요시합니다.
-      </p>
+      <p>몰입과 경험을 중요시합니다.</p>
       <p>
         익숙하지 않은 기술을 사용하는 것에 흥미를 느끼며 코드를 작성할 때 다양한 접근법을 통해 구현하는 것을 즐깁니다.
       </p>
-      <p>
-        협업 간 커뮤니케이션을 통한 공유와 짧은 간격의 피드백을 통한 발전을 추구합니다.
-      </p>
+      <p>협업 간 커뮤니케이션을 통한 공유와 짧은 간격의 피드백을 통한 발전을 추구합니다.</p>
     </ProfileDescriptionBox>
-  )
-}
+  );
+};
 
 export default ProfileDescription;
