+++ src/pages/about/index.tsx
@@ -1,12 +1,16 @@
-import AboutPage from '@/components/templates/AboutPage'
-import SeoHead from '@/components/SeoHead'
-import PageType from '@/types/page'
+import SeoHead from '@/components/SeoHead';
+import TechStacks from './components/TechStacks/TechStacks';
+import ProfileCard from './components/ProfileCard/ProfileCard';
+import PageType from '@/types/page';
 
 export default function Index() {
   return (
     <>
       <SeoHead page={PageType.Main} />
-      <AboutPage />
+      <section>
+        <ProfileCard />
+        <TechStacks />
+      </section>
     </>
-  )
+  );
 }
